+++ changelog.md
@@ -431,3 +431,6 @@ are 9001 for HSQLDB, 9092 for H2 and 1527 for Derby.
 
 ## v9.0 - BuildXXXXXXXX
 - Reintroduce the Frank-Wolfe based assignments algorithms after bug fixes
+- Replace FileUtils.copyFiles(...), FileUtils.deleteFile(...) and ProjectLocker.createLock(...) with safer versions
+- More robust Local handling at startup
+